requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=3.0.0